Moving state machine tests to test files for states
