+++ tests/isar/state_machine/states/test_resuming_return_home_state.py
@@ -1,0 +1,27 @@
+from typing import Optional, cast
+
+from isar.eventhandlers.eventhandler import EventHandlerBase, EventHandlerMapping
+from isar.state_machine.state_machine import StateMachine
+
+
+def test_transition_from_return_home_paused_to_resuming_return_home(
+    sync_state_machine: StateMachine,
+) -> None:
+    sync_state_machine.state = sync_state_machine.return_home_paused_state.name  # type: ignore
+
+    return_home_paused_state: EventHandlerBase = cast(
+        EventHandlerBase, sync_state_machine.return_home_paused_state
+    )
+    event_handler: Optional[EventHandlerMapping] = (
+        return_home_paused_state.get_event_handler_by_name("resume_return_home_event")
+    )
+
+    assert event_handler is not None
+
+    event_handler.event.trigger_event(True)
+    transition = event_handler.handler(event_handler.event)
+
+    assert transition is sync_state_machine.resume  # type: ignore
+
+    transition()
+    assert sync_state_machine.state is sync_state_machine.resuming_return_home_state.name  # type: ignore
