+++ tests/isar/state_machine/states/test_unknown_status_state.py
@@ -1,0 +1,2 @@
+def test_initial_unknown_status(state_machine) -> None:
+    assert state_machine.state == "unknown_status"
